requirements.txt


# Master Thesis AI Finance - Python Dependencies
# ==============================================

# Claude API Integration
anthropic==0.21.3

# Web Requests & Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Analysis & Visualization
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
openpyxl==3.1.2

# Configuration & File Handling
pyyaml==6.0.1

# Academic Research Tools
scholarly==1.7.11
bibtexparser==1.4.1
crossref-commons==0.0.7
arxiv==2.1.0

# Document Processing
python-docx==1.1.0

# Development Tools
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# Testing & Quality
pytest==7.4.4
black==23.12.1
flake8==7.0.0

# Backup & Utilities
schedule==1.2.0
pathlib2==2.3.7

# RAG System & Vector Database
qdrant-client==1.7.0
faiss-cpu==1.8.0
PyMuPDF==1.23.8
transformers>=4.36.0
torch==2.1.0
torchvision==0.16.0
networkx==3.2.1
sentence-transformers>=3.0.0
huggingface_hub>=0.23.0

# Database Clients
psycopg2-binary==2.9.9
redis==5.0.1
minio==7.2.0

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
langchain==0.1.0
langchain-community==0.0.10

# Additional Utilities
aiofiles==23.2.1
httpx==0.27.0
pytesseract==0.3.10
ollama==0.1.9
streamlit==1.35.0
psutil==5.9.8
watchdog==4.0.1
aiohttp>=3.9.0
